+++ apps/backend/src/modules/user/modules/secrets/interfaces/responses.ts
@@ -1,7 +1,13 @@
+import type { IUserSecretsGetAll } from './user-secrets';
+
 export interface ICreateClientSecret {
 	clientSecret: string;
 }
 
 export interface IRefreshClientSecret {
 	clientSecret: string;
+}
+
+export interface IGetAllSecretsResponse {
+	secrets: IUserSecretsGetAll[];
 }