+++ apps/backend/src/modules/user/modules/secrets/secrets.module.ts
@@ -5,6 +5,7 @@ import { CommandHandlers } from './commands/handlers';
 import { CreateController } from './create.controller';
 import { DeleteController } from './delete.controller';
 import { EditSecretController } from './edit-label.controller';
+import { GetAllController } from './get-all.controller';
 import { BelongingSecretGuard } from './guards/belonging-secret.guard';
 import { QueryHandlers } from './queries/handlers';
 import { RefreshSecretController } from './refresh-secret.controller';
@@ -12,7 +13,13 @@ import { SecretsService } from './secrets.service';
 
 @Module({
 	imports: [CqrsModule],
-	controllers: [DeleteController, RefreshSecretController, CreateController, EditSecretController],
+	controllers: [
+		DeleteController,
+		RefreshSecretController,
+		CreateController,
+		EditSecretController,
+		GetAllController,
+	],
 	providers: [...CommandHandlers, ...QueryHandlers, BelongingSecretGuard, SecretsService],
 })
 export class SecretsModule {}